[Gold I] Title: K번째 수, Time: 144 ms, Memory: 14228 KB -BaekjoonHub

+++ 백준/Gold/1300. K번째 수/K번째 수.java
@@ -6,12 +6,13 @@ class Main {
 	public static void main(String[] args) throws Exception {
 		BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 		// StringTokenizer st = new StringTokenizer(br.readLine());
-		long N = Integer.parseInt(br.readLine());
-		long k = Integer.parseInt(br.readLine());
+		long N = Long.parseLong(br.readLine());
+		long k = Long.parseLong(br.readLine());
 
 		long start = 1, end = N * N, answer = -1;
 		while (start <= end) {
 			long mid = (start + end) / 2;
+
 			long sum = 0;
 			for (int i = 1; i <= N; i++)
 				sum += Math.min(mid / i, N);
